requirements.txt
annotated-types==0.7.0
anyio==4.4.0
aws_secretsmanager_caching==1.1.3
boto3==1.35.0
botocore==1.35.0
certifi==2024.7.4
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.0
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
mangum==0.17.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==24.0
pluggy==1.5.0
psycopg2-binary ==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
rich==13.7.1
s3transfer==0.10.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.23.0
websockets==12.0
